Перенес работу с сервером в папку api
