+++ src/index.jsx
@@ -6,7 +6,7 @@ import {BrowserRouter} from 'react-router-dom';
 import {composeWithDevTools} from "redux-devtools-extension";
 import App from './components/app/app';
 import reducer from "./reducer";
-import configureAPI from './api';
+import configureAPI from './api/api';
 
 const api = configureAPI();
