--- src/reducer/data/operations.js
@@ -1,14 +1,0 @@
-import {loadFilms} from "./data";
-
-const loadFilmsRequest = () => (dispatch, _getState, api) => {
-  return api.get(`/films`)
-    .then((response) => {
-      dispatch(loadFilms(response.data));
-    });
-};
-
-export {
-  loadFilmsRequest
-};
-
-
