--- src/reducer/user/operations.js
@@ -1,15 +1,0 @@
-import {singIn} from "./user";
-
-const singInRequest = (email, password) => (dispatch, _getState, api) => {
-  return api.post(`/login`, {
-    email,
-    password
-  })
-    .then((response) => {
-      dispatch(singIn(response.data));
-    });
-};
-
-export {
-  singInRequest
-};
